requirements.txt
streamlit==1.36.0
transformers==4.38.1
greenlet
openai
python-dotenv
vertexai
